Altered motor_turn_test.c
I noticed some of the lines didn't look like what was in Assignment 5. If they were like that for a reason definitely change them back.
With the code as it is now, wheel on pins 0 and 1 should both spin at max speed in opposite directions. We will have to adjust the sleep duration though.

This should be made a method/function which can be called whenever we want the robot to turn.