Tried to add in Elf's code
I took a section of the skeleton code so that we could play around with it in this file - it almost certainly has many errors as it is now.
I'm not sure if examples on Elf's wiki are not exactly what he showed in the lecture? So we should maybe check the lecture slides and work out exactly how to do this.